requirements.txt
Flask==3.1.2
numpy==2.1.3
tensorflow==2.19.0
Pillow==12.0.0